Automate pip upload
Reorganize repository: use TOML for configuration, add GitHub action for auto-pushing to pypi

+++ src/power1d/prob.py
@@ -56,7 +56,7 @@ FOUR_LOG2   = 4*log(2)
 SQRT_4LOG2  = sqrt(4*log(2))
 SQRT_2      = sqrt(2)
 TWO_PI      = 2*pi
-eps         = np.finfo(np.float).eps
+eps         = np.finfo(float).eps
